6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long alonglong;
 	float afloat;
 
-	printf("Size of a char: %lu byte(S)\n", sizeof(acharacter));
-	printf("Size of an int: %lu byte(S)\n", sizeof(ainteger));
-	printf("Size of a long int: %lu byte(S)\n", sizeof(along));
-	printf("Size of a long long int: %lu byte(S)\n", sizeof(alonglong));
-	printf("Size of a float: %lu byte(S)\n", sizeof(afloat));
+	printf("Size of a char: %lu byte(S)\n",(unsigned long)sizeof(acharacter));
+	printf("Size of an int: %lu byte(S)\n", (unsigned long)sizeof(ainteger));
+	printf("Size of a long int: %lu byte(S)\n", (unsigned long)sizeof(along));
+	printf("Size of a long long int: %lu byte(S)\n", (unsigned long)sizeof(alonglong));
+	printf("Size of a float: %lu byte(S)\n", (unsigned long)sizeof(afloat));
 return (0);
 }
